Complete migration of event content arbs
