+++ lib-test/api/lib-test.api
@@ -2,6 +2,7 @@ public final class app/cash/nostrino/ArbPrimitive {
 	public static final field INSTANCE Lapp/cash/nostrino/ArbPrimitive;
 	public final fun getArbByteString32 ()Lio/kotest/property/Arb;
 	public final fun getArbByteString64 ()Lio/kotest/property/Arb;
+	public final fun getArbEmoji ()Lio/kotest/property/Arb;
 	public final fun getArbInstantSeconds ()Lio/kotest/property/Arb;
 	public final fun getArbUUID ()Lio/kotest/property/Arb;
 	public final fun getArbVanillaString ()Lio/kotest/property/Arb;
@@ -13,8 +14,15 @@ public final class app/cash/nostrino/crypto/ArbKeys {
 	public final fun getArbSecKey ()Lio/kotest/property/Arb;
 }
 
+public final class app/cash/nostrino/model/ArbEvent {
+	public static final field INSTANCE Lapp/cash/nostrino/model/ArbEvent;
+	public final fun getArbEventId ()Lio/kotest/property/Arb;
+}
+
 public final class app/cash/nostrino/model/ArbEventContent {
 	public static final field INSTANCE Lapp/cash/nostrino/model/ArbEventContent;
+	public final fun getArbEncryptedDm ()Lio/kotest/property/Arb;
+	public final fun getArbFilter ()Lio/kotest/property/Arb;
 	public final fun getArbReaction ()Lio/kotest/property/Arb;
 	public final fun getArbTextNote ()Lio/kotest/property/Arb;
 	public final fun getArbUserMetaData ()Lio/kotest/property/Arb;
