+++ lib/src/test/kotlin/app/cash/nostrino/model/FilterTest.kt
@@ -1,14 +1,10 @@
 package app.cash.nostrino.model
 
-import app.cash.nostrino.ArbPrimitive.arbInstantSeconds
 import app.cash.nostrino.crypto.PubKeyTest.Companion.arbPubKey
-import app.cash.nostrino.model.EventTest.Companion.arbEventId
+import app.cash.nostrino.model.ArbEventContent.arbFilter
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
 import io.kotest.property.Arb
-import io.kotest.property.arbitrary.bind
-import io.kotest.property.arbitrary.map
-import io.kotest.property.arbitrary.orNull
 import io.kotest.property.arbitrary.pair
 import io.kotest.property.arbitrary.set
 import io.kotest.property.checkAll
@@ -26,18 +22,4 @@ class FilterTest : StringSpec({
     }
   }
 
-}) {
-  companion object {
-    val arbFilter: Arb<Filter> = Arb.bind(
-      Arb.set(arbEventId).orNull(),
-      arbInstantSeconds.orNull(),
-      Arb.set(arbPubKey.map { it.key.hex() }).orNull(),
-    ) { ids, since, authors ->
-      Filter(
-        ids = ids,
-        since = since,
-        authors = authors,
-      )
-    }
-  }
-}
+})
